chore(bench): Add possibility to specify runs_on: TARGET at group: runs level
